Move to user router/services
