+++ app.js
@@ -8,7 +8,6 @@ const monitorLogRouter = require('./routes/monitorLog.router');
 const speciesRouter = require('./routes/species.router');
 const emailRouter = require('./routes/nodemailer.router');
 const adminInviteRouter = require('./routes/adminInvite.router');
-const assignmentsRouter = require('./routes/assignments.service');
 
 const userRouter = require('./routes/user.router');
 const { authRouter, verifyToken } = require('./routes/auth.router');
@@ -38,7 +37,6 @@ app.use(cookieParser());
 
 app.use('/users', userRouter);
 app.use('/auth', authRouter);
-app.use('/assigned', assignmentsRouter);
 // TODO remove
 app.use('/test', [verifyToken, userRouter]);
