+++ routes/user.router.js
@@ -72,6 +72,32 @@ router.get('/', async (req, res) => {
   }
 });
 
+// get user's assigned segments
+// returns [ {Segment}, {Segment} ]
+router.get('/segments/:id', async (req, res) => {
+  const { id } = req.params;
+  try {
+    const assignedSegments = await userService.getAssignedSegments(id);
+    res.status(200).send(assignedSegments);
+  } catch (err) {
+    console.error(err);
+    res.send(400).json({ message: err.message });
+  }
+});
+
+// get all of user's submitted monitor logs
+// returns [ { userSubmissions: [Submission, Submission, ...] } ]
+router.get('/submissions/:id', async (req, res) => {
+  const { id } = req.params;
+  try {
+    const userSubmissions = await userService.getUserSubmissions(id);
+    res.status(200).send(userSubmissions);
+  } catch (err) {
+    console.error(err);
+    res.send(400).json({ message: err.message });
+  }
+});
+
 // delete profile
 router.delete('/:id', async (req, res) => {
   const { id } = req.params;
